+++ example/custom_index_name.dart
@@ -1,8 +1,8 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-library example;
+library custom_library_name;
 
-export 'custom_index_name.dart';
+export 'example.dart';
 export 'export_folders.dart';
 export 'index_generator_example.dart';
 export 'pick.dart';