+++ example/export_folders.dart
@@ -1,0 +1,5 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+library example;
+
+export 'src/pick.dart';