+++ example/index_generator_example.dart
@@ -1,7 +1,6 @@
 import 'package:index_generator/src/index_generator.dart';
 
-import 'file:///Users/sbressan/Documents/projects/dart_flutter_packages/index_generator/bin/main.dart'
-    as generator;
+import '../bin/main.dart' as generator;
 
 void main() {
   isTestingMode = true;